+++ src/MOGWAI/MOGWAI/Engine/Error.cs
@@ -20,11 +20,11 @@ namespace MOGWAI.Engine
 
         public static readonly Error None = RegisterError("MW.0", "OK");
         public static readonly Error ParseError = RegisterError("MW.1", "parse error");
-        public static readonly Error HaltEncountedError = RegisterError("MW.2", "halt encounted error");
+        public static readonly Error HaltEncounteredError = RegisterError("MW.2", "halt encountered error");
         public static readonly Error EmptyCodeError = RegisterError("MW.3", "empty code error");
         public static readonly Error InternalError = RegisterError("MW.4", "internal error");
         public static readonly Error PlatformNotSupportedError = RegisterError("MW.5", "platform not supported error");
-        public static readonly Error UnabledToFireEventError = RegisterError("MW.6", "unabled to fire event error");
+        public static readonly Error UnableToFireEventError = RegisterError("MW.6", "unable to fire event error");
         public static readonly Error OperationNotSupportedError = RegisterError("MW.7", "operation not supported error");
         public static readonly Error CircularReferenceError = RegisterError("MW.8", "circular reference error");
         public static readonly Error AssertError = RegisterError("MW.9", "assert error");
@@ -43,23 +43,23 @@ namespace MOGWAI.Engine
         public static readonly Error ConvertError = RegisterError("MW.32", "convert error");
 
         public static readonly Error UnknownNameError = RegisterError("MW.40", "unknown name error");
-        public static readonly Error NameAlreadyExistsError = RegisterError("MW.41", "name already exits error");
+        public static readonly Error NameAlreadyExistsError = RegisterError("MW.41", "name already exists error");
         public static readonly Error FunctionAlreadyExistsError = RegisterError("MW.42", "function already exists error");
         public static readonly Error NameAlreadyUsedByFunctionError = RegisterError("MW.43", "name already used by function error");
         public static readonly Error NameAlreadyUsedByVarError = RegisterError("MW.44", "name already used by var error");
         public static readonly Error UnknownKeyError = RegisterError("MW.45", "unknown key error");
         public static readonly Error InvalidNameError = RegisterError("MW.46", "invalid name error");
-        public static readonly Error UnabledToWriteValueError = RegisterError("MW.47", "unabled to write value in var");
-        public static readonly Error UnabledToWriteValueInUndeclaredVarError = RegisterError("MW.48", "unabled to write value in undeclared var");
+        public static readonly Error UnableToWriteValueError = RegisterError("MW.47", "unable to write value in var error");
+        public static readonly Error UnableToWriteValueInUndeclaredVarError = RegisterError("MW.48", "unable to write value in undeclared var error");
 
         public static readonly Error UnknownWordError = RegisterError("MW.50", "unknown word error");
 
         public static readonly Error TaskCreationError = RegisterError("MW.60", "task creation error");
-        public static readonly Error UnabledToStartTaskError = RegisterError("MW.61", "unabled to start task error");
-        public static readonly Error InvalidOutsideOfaTaskError = RegisterError("MW.62", "invalid outside of a task error");
+        public static readonly Error UnabledToStartTaskError = RegisterError("MW.61", "unable to start task error");
+        public static readonly Error InvalidOutsideOfATaskError = RegisterError("MW.62", "invalid outside of a task error");
 
         public static readonly Error InvalidPathError = RegisterError("MW.70", "invalid path error");
-        public static readonly Error PathDoesNotExistsError = RegisterError("MW.71", "path does not exists error");
+        public static readonly Error PathDoesNotExistError = RegisterError("MW.71", "path does not exist error");
         public static readonly Error FileOperationError = RegisterError("MW.72", "file operation error");
         public static readonly Error UnknownFileError = RegisterError("MW.73", "unknown file error");