+++ src/MOGWAI/MOGWAI/Primitives/PrimitiveFOREVER.cs
@@ -55,7 +55,7 @@ namespace MOGWAI.Primitives
 
                     if (Engine.HaltRequested)
                     {
-                        result = EvalResult.Failure(Engine, Error.HaltEncountedError);
+                        result = EvalResult.Failure(Engine, Error.HaltEncounteredError);
                         break;
                     }